rename deploment to web-deploy and add web-deploy in schame.json
