+++ apps/cli/src/helpers/project-generation/command-handlers.ts
@@ -5,8 +5,8 @@ import pc from "picocolors";
 import { DEFAULT_CONFIG } from "../../constants";
 import { getAddonsToAdd } from "../../prompts/addons";
 import { gatherConfig } from "../../prompts/config-prompts";
-import { getDeploymentToAdd } from "../../prompts/deployment";
 import { getProjectName } from "../../prompts/project-name";
+import { getDeploymentToAdd } from "../../prompts/web-deploy";
 import type { AddInput, CreateInput, ProjectConfig } from "../../types";
 import { trackProjectCreation } from "../../utils/analytics";
 import { displayConfig } from "../../utils/display-config";
