+++ apps/web/public/analytics-data.json
@@ -140013,6 +140013,6 @@
     }
   ],
   "lastUpdated": "Jul 1, 2025, 03:45 AM",
-  "generatedAt": "2025-07-04T18:27:32.233Z",
+  "generatedAt": "2025-07-05T09:58:58.586Z",
   "totalRecords": 5637
 }
